// src/member/member-pelkat.enum.ts
export enum MemberPelkat {
  PELAYANAN_ANAK = 'Pelayanan Anak',
  PERSEKUTUAN_TARUNA = 'Persekutuan Taruna',
  GERAKAN_PEMUDA = 'Gerakan Pemuda',
  PERSEKUTUAN_KAUM_BAPAK = 'Persekutuan Kaum Bapak',
  PERSEKUTUAN_KAUM_PEREMPUAN = 'Persekutuan Kaum Perempuan',
  PERSEKUTUAN_KAUM_LANJUT_USIA = 'Persekutuan Kaum Lanjut Usia',
}